HDPI hybrid adjustments
With com.android.systemui.statusbar.dpi set at 240 there's an unsightly gap below the status bar
http://db.tt/MIBksfk2
With the adjustments a statusbar setting of 160-240 looks proper
http://db.tt/F2xnGOM8
The Navbar was also huge. I have it set to 0 because it's unneeded on many hdpi devices. If, it's desirable to keep it, I suggest a number between 75-100%